mood_board/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==3.3
pip==21.3.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
setuptools==59.1.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.0
urllib3==1.26.7
wheel==0.37.0
